+++ BlazorStudio.ClassLib/Store/PlainTextEditorCase/PlainTextEditorStates.PlainTextEditorRecord.cs
@@ -45,7 +45,7 @@ public partial record PlainTextEditorStates
         public int RowIndexOffset { get; init; }
         public int CharacterIndexOffsetRelativeToRow { get; init; }
         public int CurrentCharacterColumnIndex { get; init; }
-        public int CurrentPositionIndex { get; init; }
+        public long CurrentPositionIndex { get; init; }
         public int PreviouslySetCharacterColumnIndex { get; init; }
         public int CharacterColumnIndexOffset { get; init; }
         public SelectionSpanRecord? SelectionSpan { get; init; }
